CONFIG-1987: move log and meta to seprate file
