+++ query/src/configs/mod.rs
@@ -18,13 +18,16 @@ mod macros;
 #[cfg(test)]
 mod config_test;
 
-pub mod config;
+mod config;
+mod config_log;
+mod config_meta;
 mod config_query;
 mod config_storage;
 
 pub use common_store_api_sdk::RpcClientTlsConfig;
 pub use config::Config;
-pub use config::LogConfig;
-pub use config::MetaConfig;
-pub use config_query::*;
-pub use config_storage::*;
+pub use config::DATABEND_COMMIT_VERSION;
+pub use config_log::LogConfig;
+pub use config_meta::MetaConfig;
+pub use config_query::QueryConfig;
+pub use config_storage::StorageConfig;
